+++ pyng/space/phys_obj.py
@@ -57,8 +57,8 @@ class Point(PhysObj):
 
     def is_inside_of_other_object(self, other_object) -> bool:
         return (
-                self.position.x == other_object.position.x
-                and self.position.y == other_object.position.y
+            self.position.x == other_object.position.x
+            and self.position.y == other_object.position.y
         )
 
 
@@ -105,10 +105,15 @@ class ConvexPolygon(PhysObj):
         return [(min_x, min_y), (max_x, max_y)]
 
     def is_axis_aligned_rectangle(self) -> bool:
-        """ Återvänder sant om objektet är en rektangel som inte har roterats """
+        """Återvänder sant om objektet är en rektangel som inte har roterats"""
         # TODO: Kolla om bättre implementation finns när det kommer till
         #       om rektangeln roterats
-        non_rotated_angles = [math.pi / 4, 3 * math.pi / 4, 5 * math.pi / 4, 7 * math.pi / 4]
+        non_rotated_angles = [
+            math.pi / 4,
+            3 * math.pi / 4,
+            5 * math.pi / 4,
+            7 * math.pi / 4,
+        ]
         return self.num_of_sides == 4 and self.angle in non_rotated_angles
 
     def check_axis_aligned_collision(self, other_rect) -> bool:
@@ -125,10 +130,17 @@ class ConvexPolygon(PhysObj):
         y_min = self.position.y - side_len
         other_y_max = other_rect.position.y + other_side_len
         other_y_min = other_rect.position.y - other_side_len
-        return x_min < other_x_max and other_x_min < x_max and y_min < other_y_max and other_y_min < y_max
+        return (
+            x_min < other_x_max
+            and other_x_min < x_max
+            and y_min < other_y_max
+            and other_y_min < y_max
+        )
 
     def is_inside_of(self, other_rect):
-        if self.is_axis_aligned_rectangle() and other_rect.is_axis_aligned_rectangle():  # Använd AABB
+        if (
+            self.is_axis_aligned_rectangle() and other_rect.is_axis_aligned_rectangle()
+        ):  # Använd AABB
             return self.check_axis_aligned_collision(other_rect)
 
     def render(self, view_model):
@@ -154,54 +166,12 @@ class Circle(PhysObj):
 
     def is_inside_of(self, other) -> bool:
         if isinstance(other, Circle):
-            return self.position.distance_to(other.position) < self.radius + other.radius
+            return (
+                self.position.distance_to(other.position) < self.radius + other.radius
+            )
 
 
 class CollisionResult:
     def __init__(self, contact_point: Vector2D, normal: Vector2D):
         self.contact_point = contact_point
         self.normal = normal
-
-# def RayVsRect(ray_origin: Vector2D, ray_direction: Vector2D, target: Rectangle) -> bool:
-#     t_near = (target.position - ray_origin).element_division(ray_direction)
-#     t_far = (
-#         target.position + Vector2D(target.width, target.height) - ray_origin
-#     ).element_division(ray_direction)
-#     if t_near.x > t_far.x:
-#         t_near.x, t_far.x = t_far.x, t_near.x
-
-#     if t_near.y > t_far.y:
-#         t_near.y, t_far.y = t_far.y, t_near.y
-
-#     # Nx < Fy eller Ny < Fx
-#     if t_near.x > t_far.y or t_near.y < t_far.x:
-#         return False
-
-#     t_hit_near = max(t_near.x, t_near.y)
-#     t_hit_far = min(t_far.x, t_far.y)
-
-#     if t_hit_far < 0:
-#         return False
-
-#     contact_point = ray_origin + (t_hit_near * ray_direction)
-
-#     # Hitta vilken sida som träffas
-
-#     if t_near.x > t_near.y:
-#         if ray_direction.x < 0:
-#             normal = Vector2D(1, 0)
-#         else:
-#             normal = Vector2D(-1, 0)
-
-#     elif t_near.x < t_near.y:
-#         if ray_direction.y < 0:
-#             normal = Vector2D(0, 1)
-#         else:
-#             normal = Vector2D(0, -1)
-
-#     return True, CollisionResult(contact_point, normal)
-
-#     # TODO: FINISH ME
-
-
-# RayVsRect(Vector2D(0, 0), Vector2D(1, 1), Rectangle(1, RED, 1, 1))
